chore: rename adapters/extraction/ into adapters/extractors/
