+++ libs/core/kiln_ai/adapters/__init__.py
@@ -19,7 +19,7 @@ The eval submodule contains the code for evaluating the performance of a model.
 from . import (
     data_gen,
     eval,
-    extraction,
+    extractors,
     fine_tune,
     ml_model_list,
     model_adapters,
@@ -35,5 +35,5 @@ __all__ = [
     "prompt_builders",
     "repair",
     "eval",
-    "extraction",
+    "extractors",
 ]
