+++ libs/core/kiln_ai/adapters/extractors/base_extractor.py
